Moved unit tests, updated version.
